Move more stuff to ecs_helper.py
Also, damn Python imports :-)
