+++ scripts/generators/ecs_helpers.py
@@ -13,6 +13,12 @@ def dict_copy_keys_ordered(dict, copied_keys):
     return ordered_dict
 
 
+def dict_copy_existing_keys(source, destination, keys):
+    for key in keys:
+        if key in source:
+            destination[key] = source[key]
+
+
 def yaml_ordereddict(dumper, data):
     # YAML representation of an OrderedDict will be like a dictionary, but
     # respecting the order of the dictionary.
@@ -23,6 +29,7 @@ def yaml_ordereddict(dumper, data):
         node_value = dumper.represent_data(item_value)
         value.append((node_key, node_value))
     return yaml.nodes.MappingNode(u'tag:yaml.org,2002:map', value)
+
 
 yaml.add_representer(OrderedDict, yaml_ordereddict)
