+++ scripts/generators/intermediate_files.py
@@ -1,3 +1,6 @@
+from generators import ecs_helpers
+
+
 def generate(ecs_nested, ecs_flat):
     ecs_helpers.yaml_dump('generated/ecs/fields_flat.yml', ecs_flat)
     ecs_helpers.yaml_dump('generated/ecs/fields_nested.yml', ecs_nested)
