+++ scripts/tests/test_asciidoc_fields.py
@@ -1,4 +1,9 @@
+import os
+import sys
 import unittest
+
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
 from scripts.generators import asciidoc_fields
