+++ scripts/tests/test_ecs_spec.py
@@ -1,5 +1,11 @@
+import os
+import sys
 import unittest
+
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
 from scripts import schema_reader
+
 
 (nested, flat) = schema_reader.load_ecs()
