+++ scripts/tests/test_es_template.py
@@ -1,4 +1,9 @@
+import os
+import sys
 import unittest
+
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
 from scripts.generators import es_template
 
 
@@ -24,20 +29,6 @@ class TestGeneratorsEsTemplate(unittest.TestCase):
             'key2': {'properties': {'nested1': 'value12', 'nested2': 'value22'}},
             'key3': 'value3'
         })
-
-    # dict_copy_existing_keys
-
-    def test_dict_copy_existing_keys(self):
-        source = {'key1': 'value1'}
-        destination = {}
-        es_template.dict_copy_existing_keys(source, destination, ['key1', 'missingkey'])
-        self.assertEqual(destination, {'key1': 'value1'})
-
-    def test_dict_copy_existing_keys_overwrites(self):
-        source = {'key1': 'new_value'}
-        destination = {'key1': 'overwritten', 'untouched': 'untouched'}
-        es_template.dict_copy_existing_keys(source, destination, ['key1', 'untouched'])
-        self.assertEqual(destination, {'key1': 'new_value', 'untouched': 'untouched'})
 
 
 if __name__ == '__main__':
