+++ scripts/tests/test_schema_reader.py
@@ -1,4 +1,9 @@
+import os
+import sys
 import unittest
+
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
 from scripts import schema_reader
